Locking down deleteItem permissions

+++ src/resolvers/Mutation.js
@@ -46,7 +46,16 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // find the item
-    const item = await ctx.db.query.item({ where }, `{id title}`);
+    const item = await ctx.db.query.item({ where }, `{id title user { id }}`);
+    // check if owner or user has permissions
+    const owner = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission),
+    );
+
+    if (!owner && !hasPermissions) {
+      throw new Error("You don't have permissions to delete this item!");
+    }
     // delete it
     return ctx.db.mutation.deleteItem({ where }, info);
   },
